Add spatie permissions

+++ app/Livewire/Category/Upsert.php
@@ -5,6 +5,7 @@ namespace App\Livewire\Category;
 use App\Livewire\Utils\Modal;
 use App\Livewire\Utils\Slug;
 use App\Models\Category;
+use Illuminate\Support\Facades\Gate;
 use Livewire\Component;
 use Livewire\Attributes\On;
 use Masmerise\Toaster\Toaster;
@@ -39,10 +40,14 @@ class Upsert extends Component
     public function openModal(?int $id = null): void
     {
         if ($id) {
+            abort_if(Gate::denies('edit_category'), 403);
+
             $this->category = Category::findOrFail($id);
             $this->title = $this->category->title;
             $this->slug = $this->category->slug;
         } else {
+            abort_if(Gate::denies('create_category'), 403);
+
             $this->resetModal($this->fields);
         }
 
@@ -56,9 +61,13 @@ class Upsert extends Component
             'slug' => 'required|string|min:4|max:255',
         ];
 
-        // Add unique rules if editing existing category.
         if ($this->category) {
+            abort_if(Gate::denies('edit_category'), 403);
+
             $rules['slug'] .= '|unique:categories,slug,' . $this->category->id;
+        }
+        else {
+            abort_if(Gate::denies('create_category'), 403);
         }
 
         $validated = $this->validate($rules);
@@ -80,6 +89,8 @@ class Upsert extends Component
     #[On('do-category-delete')]
     public function deleteModal(?int $id = null): void
     {
+        abort_if(Gate::denies('delete_category'), 403);
+
         Category::findOrFail($id)->delete();
         Toaster::success(__('Category deleted successfully!'));
